+++ R/eta_pairs.R
@@ -1,11 +1,20 @@
-eta_pairs_fun <- function(data, mapping, ...) {
+pairs_lower <- function(data, mapping, smooth_color = .ggblue, smooth_lty = 2, ...) {
+
+  if(is.character(mapping$smooth_color)){
+    smooth_color <- mapping$smooth_color
+  }
+  if(is.numeric(mapping$smooth_lty)) {
+    smooth_lty <- mapping$smooth_lty
+  }
+
   ggplot(data = data, mapping = mapping) +
     geom_point() +
-    geom_smooth(method = "loess", color = .ggblue, lty=2, se = FALSE, lwd = 1.3)
+    geom_smooth(method = "loess", color = smooth_color, lty = smooth_lty,
+                se = FALSE, lwd = 1.3)
 
 }
 
-eta_pairs_upper <- function(data, mapping, shk = list(), ...) {
+pairs_upper <- function(data, mapping, ...) {
   x <- deparse(mapping$x)[1]
   y <- deparse(mapping$y)[1]
   label <- as.character(signif(cor(data[,x],data[,y],use = "complete.obs"), digits=3))
@@ -18,7 +27,7 @@ eta_pairs_upper <- function(data, mapping, shk = list(), ...) {
 
 
 
-##' Pairs plot for ETAs using ggpairs
+##' Pairs plots using ggpairs
 ##'
 ##'
 ##'
@@ -28,33 +37,35 @@ eta_pairs_upper <- function(data, mapping, shk = list(), ...) {
 ##' @param alpha passed to \code{geom_histogram}
 ##' @param fill passed to \code{geom_histogram}
 ##' @param col passed to \code{geom_histogram}
-##' @param shk not used
+##' @param upper_fun function to use for \code{upper} argument
+##' @param lower_fun function to use for \code{lower} argument
+##' @param ... passed to \code{GGally::ggpairs}
 ##'
 ##' @details
-##' This funciton requires the \code{GGally} package to be installed.
+##' This function requires the \code{GGally} package to be installed.
 ##'
 ##' @return
 ##' The result from a ggpairs call.
 ##'
 ##' @examples
 ##'
-##' library(dplyr)
+##' df <- dplyr::filter(pmplots_data(), EVID==0)
+##' id <- dplyr::distinct(df, ID, .keep_all = TRUE)
 ##'
-##' id <- pmplots_data() %>%
-##'   filter(EVID==0) %>%
-##'   distinct(ID, .keep_all = TRUE)
 ##'
 ##' eta_pairs(id, c("ETA1//ETA-CL", "ETA2//ETA-VC", "ETA3//ETA-KA"))
 ##'
 ##' @export
-eta_pairs <- function(x, etas, bins = 15, alpha = 0.6, fill = "black",
-                      col="grey", shk = list()) {
+pairs_plot <- function(x, etas, bins = 15, alpha = 0.6, fill = "black",
+                      col="grey",
+                      upper_fun = pairs_upper, lower_fun = pairs_lower, ...) {
 
   if(!requireNamespace("GGally")) {
     stop("this function requires that the GGally package be installed",
          call. = FALSE)
   }
-  diag <- GGally::wrap("barDiag", bins = bins, alpha = alpha, fill=fill, col=col)
+  diag <- GGally::wrap("barDiag", bins = bins,
+                       alpha = alpha, fill=fill, col=col)
   x <- as.data.frame(x)
   etal <- lapply(etas, col_label)
   cols <- sapply(etal, "[[", 1L)
@@ -64,9 +75,16 @@ eta_pairs <- function(x, etas, bins = 15, alpha = 0.6, fill = "black",
   for(col in cols) {
     require_numeric(x,col)
   }
-  GGally::ggpairs(x[,cols],
+  GGally::ggpairs(x, ...,
+                  columns=cols,
                   columnLabels=labs,
-                  upper = list(continuous = eta_pairs_upper, shk = shk),
+                  upper = list(continuous = pairs_upper),
                   diag = list(continuous = diag),
-                  lower = list(continuous = eta_pairs_fun))
+                  lower = list(continuous = pairs_lower)) + pm_theme()
+}
+
+##' @rdname pairs_plot
+##' @export
+eta_pairs <- function(...) {
+  pairs_plot(...)
 }