+++ R/hist.R
@@ -1,37 +1,39 @@
-
-##' ETA histograms
+##' Generate a histogram plot
 ##'
-##' @param df data frame to plot
-##' @param x character col//label for x-axis data; see \code{\link{col_label}}
-##' @param xs see \code{\link{defx}}
-##' @param fill passed to \code{geom_histogram}
-##' @param col passed to \code{geom_histogram}
-##' @param alpha passed to \code{geom_histogram}
-##' @param ... other arguments for \code{geom_histogram}
+##' \code{cont_hist_list} is a vectorized version
+##' of \code{cont_hist}.
 ##'
-##' @details
-##' The \code{x} column must be numeric.
+##' @param df the data frame containing plotting data
+##' @param x the x column
+##' @param xs a list of information for the x axis
+##' @param fill a character value passed to \code{geom_histogram}
+##' @param col a character value passed to \code{geom_histogram}
+##' @param alpha a numeric value passed to \code{geom_histogram}
+##' @param ... passed to \code{geom_histogram}
 ##'
-##' @examples
 ##'
-##' df <- dplyr::filter(pmplots_data(), EVID==0)
-##'
-##' etas <- c("ETA1//ETA-CL", "ETA2//ETA-V2", "ETA3//ETA-KA")
-##'
-##' eta_hist(df, etas)
 ##'
 ##' @export
-eta_hist <- function(df, x, xs=defx(), fill="black", col="white",
-                     alpha=0.6, ...) {
-  out <- vector(mode="list", length=length(x))
-  for(i in seq_along(x)) {
-    xx <- col_label(x[i])
-    require_numeric(df,xx[1])
-    xs$name <- xx[2]
-    xscale <- do.call("scale_x_continuous", xs)
-    out[[i]] <- ggplot(data=df, aes_string(x=xx[1])) +
-      geom_histogram(...,col=col,fill=fill,alpha=alpha) +
-      xscale
-  }
-  return(out)
+cont_hist <- function(df, x, xs = defx(), fill = "black",
+                      col = "white", alpha = 0.6, ...) {
+  xscale <- do.call("scale_x_continuous", xs)
+  xx <- col_label(x)
+  xscale$name <- xx[2]
+  require_numeric(df,xx[1])
+  ggplot(data=df, aes_string(x = xx[1])) +
+    pm_histogram(..., col = col, fill = fill, alpha = alpha) +
+    xscale + pm_theme()
+}
+
+##' @rdname cont_hist
+##' @export
+cont_hist_list <- function(df, x, ...) {
+  list_plot_x(df, x, ...)
+}
+
+##' @rdname cont_hist
+##' @export
+pm_histogram <- function(... , col = "white", fill = "black",
+                         alpha = 0.6) {
+  geom_histogram(..., col = col, fill = fill, alpha = alpha)
 }