+++ R/qq.R
@@ -8,9 +8,6 @@ qq_reg_data <- function(y) {
   c(int = int, slope = slope)
 }
 
-
-
-
 ##' QQ plot for conditional residuals
 ##'
 ##' @param df data frame to plot
@@ -47,7 +44,7 @@ wres_q <- function(df, x="WRES", xs = defx(), ys=defy(), abline=NULL, ...) {
   if(!is.null(abline)) {
     p <- p + geom_abline(intercept=abline[1], slope=abline[2])
   }
-  p
+  p + pm_theme()
 }
 
 ##' @export